fix: Priority local speed limit

+++ api/proxypanel/proxypanel.go
@@ -592,7 +592,7 @@ func (c *APIClient) ParseSSUserListResponse(userInfoResponse *json.RawMessage) (
 		if c.SpeedLimit > 0 {
 			speedlimit = uint64((c.SpeedLimit * 1000000) / 8)
 		} else {
-			speedlimit = uint64((user.SpeedLimit * 1000000) / 8)
+			speedlimit = user.SpeedLimit * 1000000 / 8
 		}
 		userList[i] = api.UserInfo{
 			UID:         user.UID,
